Adding class KStarsDateTime (KSDT), which is derived from ExtDateTime, and adds some astronomy-specific functionality. One important change is that KSDT stores the Julian Day internally as a long double (so it includes the time as well).  The accessor for this is "djd()", because "jd()" already exists in ExtDateTime (it returns the JD as an integer).  AFAIK, it is not possible to overload a function such that it returns a different data type.  If it is possible, please let me know.
Other important functions in KSDT:

+ ctors which take: int, double, long double (each specify JD)
                    KSDT (copy ctor)
                    ExtDate and QTime

+ setDJD(long double): sets the JD, sets the calendar date
                       and time as well.

+ setDate( ExtDate ):  Reset the date (keeps current time)
+ setTime( QTime ):  Reset the time (keeps current date)

+ comparison operators ( ==, !=, >, >=, <, <= )

+ currentDateTime() [static]: returns KSDT from CPU clock

+ jdFrac(): returns non-integer portion of djd()
+ JDat0hUT: returns JD for the current date at midnight
              (so ends in .5)

+ gst(): returns dms for the current Greenwich Sidereal Time
+ GSTtoUT( dms gst ): returns QTime for the UT corresponding
                      to gst on the date of the parent KSDT object.

+ epoch(): returns the epoch of the KSDT object
          (i.e., J2000 = 2000.0)
+ setFromEpoch( double e ):  Set the date/time from the epoch


In addition, I added some conversion functions to GeoLocation:

+ GSTtoLST( dms gst ): return the local sidereal time
                       (dms object) for the location, given gst
+ LSTtoGST( dms lst ): vice versa
+ UTtoLT( KSDT ut ):   return the local time (KSDT) for the
                       location, given the UT
+ LTtoUT( KSDT lt ):   vice versa.

So, for example, to compute LST we used to do:
    dms LST = KSUtils::UTtoLST( ut.time(), geo->lng() );

and we now do:
    dms LST = geo->GSTtoLST( ut.gst() );

In fact, all time-related functions in KSUtils have been replaced
by the new functions in KSDT or GeoLocation.  KSUtils has only
2 functions in it now.

I also tweaked the timesteps in TimeSpinBox; the large timesteps
are now multiples of a sidereal year.  This makes the animations
smoother.


unrelated fixes:

KSUtils::lagrangeInterpolation now uses long doubles instead of
doubles, so we should now have sub-second precision in determining
the exact moments of equinoxes and solstices.

Fixed memory leaks in modcalcplanets

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=311959

+++ kstars/tools/modcalcequinox.h
@@ -50,7 +50,7 @@ private:
 
 	int getYear (QString eName);
 	void showCurrentYear (void);
-	void showStartDateTime(double jd);
+	void showStartDateTime(long double jd);
 	void showSeasonDuration(float deltaJd);
 	void processLines( QTextStream &istream );
